[chore](fe-core) relocate iceberg SDK catalog tests to fe-connector-iceberg
Move the 5 iceberg catalog test classes out of fe-core's test tree into
fe-connector-iceberg (package org.apache.doris.connector.iceberg.catalog).
These tests import zero Doris classes and drive the iceberg SDK
(GlueCatalog / RESTCatalog / S3TablesCatalog) directly against real external
services; they belong with the iceberg connector, whose test classpath already
carries iceberg-core / iceberg-aws / s3-tables-catalog / junit-jupiter. This is
the last thing pinning the iceberg dependency cluster to fe-core's test compile.

  - AWSTest / IcebergGlueRestCatalogTest / IcebergUnityCatalogRestCatalogTest /
    IcebergDlfRestCatalogTest / S3TablesTest -> connector test tree, package
    updated. REST/Unity/Dlf tests move unchanged (plain guava + hadoop-common
    are already on the connector's classpath).
  - AWSTest: the connector deliberately carries only AWS SDK v2, so drop the
    non-iceberg testAWSS3 method (a bare AWS SDK v1 com.amazonaws S3 bucket-list
    smoke test) and replace the one v1 credentials-provider class literal in
    testGlueCatalog with the equivalent config string, instead of pulling the
    legacy v1 SDK onto the connector test classpath.

Verified: fe-connector-iceberg and fe-core both test-compile green.
