+++ fe/fe-connector/fe-connector-iceberg/src/test/java/org/apache/doris/connector/iceberg/catalog/AWSTest.java
@@ -15,12 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.datasource.property.metastore;
+package org.apache.doris.connector.iceberg.catalog;
 
-import com.amazonaws.auth.SystemPropertiesCredentialsProvider;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.amazonaws.services.s3.model.ObjectListing;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.LocatedFileStatus;
@@ -41,7 +37,6 @@ public class AWSTest {
     private static final String AWS_ACCESS_KEY_ID = "YOUR_ACCESS_KEY_ID"; // Replace with actual access key
     private static final String AWS_SECRET_ACCESS_KEY = "YOUR_SECRET_ACCESS_KEY"; // Replace with actual secret key
     private static final String AWS_REGION = "ap-northeast-1"; // Replace with actual region
-    private static final String S3_BUCKET_NAME = "test"; // Replace with actual bucket name
     private static final String GLUE_CATALOG_NAME = "test"; // Replace with actual catalog name
     private static final String S3A_PATH = "s3a://aws-glue-assets-123-ap-southeast-1/"; // Replace with actual S3A path
 
@@ -51,25 +46,6 @@ public class AWSTest {
         System.setProperty("aws.accessKeyId", AWS_ACCESS_KEY_ID);
         System.setProperty("aws.secretKey", AWS_SECRET_ACCESS_KEY);
         System.setProperty("aws.region", AWS_REGION);
-    }
-
-    @Test
-    public void testAWSS3() throws IOException {
-        // Create S3 client
-        AmazonS3 s3Client = AmazonS3ClientBuilder.standard()
-                .withRegion(AWS_REGION) // Set the region
-                .build();
-
-        // List S3 buckets
-        s3Client.listBuckets().forEach(bucket -> {
-            System.out.println("Bucket Name: " + bucket.getName());
-        });
-
-        // List objects in the specified S3 bucket
-        ObjectListing list = s3Client.listObjects(S3_BUCKET_NAME, "");
-        list.getObjectSummaries().forEach(objectSummary -> {
-            System.out.println("Object Key: " + objectSummary.getKey());
-        });
     }
 
     @Test
@@ -86,7 +62,7 @@ public class AWSTest {
 
         // Configure Hadoop FileSystem to use S3A with SystemPropertiesCredentialsProvider
         Configuration conf = new Configuration();
-        conf.set("fs.s3a.aws.credentials.provider", SystemPropertiesCredentialsProvider.class.getName()); // Use SystemPropertiesCredentialsProvider
+        conf.set("fs.s3a.aws.credentials.provider", "com.amazonaws.auth.SystemPropertiesCredentialsProvider");
         conf.set("fs.defaultFS", S3A_PATH);
         conf.set("fs.s3a.impl", "org.apache.hadoop.fs.s3a.S3AFileSystem");
