+++ fe/fe-connector/fe-connector-iceberg/src/test/java/org/apache/doris/connector/iceberg/catalog/IcebergGlueRestCatalogTest.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.datasource.property.metastore;
+package org.apache.doris.connector.iceberg.catalog;
 
 import com.google.common.collect.Maps;
 import org.apache.hadoop.conf.Configuration;
@@ -23,6 +23,8 @@ import org.apache.iceberg.CatalogProperties;
 import org.apache.iceberg.CatalogUtil;
 import org.apache.iceberg.FileScanTask;
 import org.apache.iceberg.Table;
+import org.apache.iceberg.aws.AwsClientProperties;
+import org.apache.iceberg.aws.s3.S3FileIOProperties;
 import org.apache.iceberg.catalog.Catalog;
 import org.apache.iceberg.catalog.Namespace;
 import org.apache.iceberg.catalog.SupportsNamespaces;
@@ -33,37 +35,28 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Map;
 
-@Disabled("set local DLF test credentials with -Ddoris.test.iceberg.dlf.access-key-id and "
-        + "-Ddoris.test.iceberg.dlf.secret-access-key before running the test")
-public class IcebergDlfRestCatalogTest {
+@Disabled("set your AWS access key, secret key and account ID before running the test")
+public class IcebergGlueRestCatalogTest {
 
-    private static final String DLF_ACCESS_KEY_ID_PROPERTY = "doris.test.iceberg.dlf.access-key-id";
-    private static final String DLF_SECRET_ACCESS_KEY_PROPERTY = "doris.test.iceberg.dlf.secret-access-key";
-    private static final String DLF_REST_URI = "http://cn-beijing-vpc.dlf.aliyuncs.com/iceberg";
-    private static final String DLF_WAREHOUSE = "new_dlf_iceberg_catalog";
-    private static final String DLF_SIGNING_NAME = "DlfNext";
-    private static final String DLF_SIGNING_REGION = "cn-beijing";
-
-    private static final String ACCESS_KEY_ID = System.getProperty(
-            DLF_ACCESS_KEY_ID_PROPERTY, "<your-access-key-id>");
-    private static final String SECRET_ACCESS_KEY = System.getProperty(
-            DLF_SECRET_ACCESS_KEY_PROPERTY, "<your-secret-access-key>");
+    private String awsAk = "";
+    private String awsSk = "";
+    private String awsAccountId = "";
 
     @Test
-    public void testIcebergDlfRestCatalog() {
-        Catalog dlfRestCatalog = initIcebergDlfRestCatalog();
-        SupportsNamespaces nsCatalog = (SupportsNamespaces) dlfRestCatalog;
+    public void testIcebergGlueRestCatalog() {
+        Catalog glueRestCatalog = initIcebergGlueRestCatalog();
+        SupportsNamespaces nsCatalog = (SupportsNamespaces) glueRestCatalog;
         // List namespaces and assert
         nsCatalog.listNamespaces(Namespace.empty()).forEach(namespace1 -> {
             System.out.println("Namespace: " + namespace1);
             Assertions.assertNotNull(namespace1, "Namespace should not be null");
 
-            dlfRestCatalog.listTables(namespace1).forEach(tableIdentifier -> {
+            glueRestCatalog.listTables(namespace1).forEach(tableIdentifier -> {
                 System.out.println("Table: " + tableIdentifier.name());
                 Assertions.assertNotNull(tableIdentifier, "TableIdentifier should not be null");
 
                 // Load table history and assert
-                Table iceTable = dlfRestCatalog.loadTable(tableIdentifier);
+                Table iceTable = glueRestCatalog.loadTable(tableIdentifier);
                 iceTable.history().forEach(snapshot -> {
                     System.out.println("Snapshot: " + snapshot);
                     Assertions.assertNotNull(snapshot, "Snapshot should not be null");
@@ -78,12 +71,14 @@ public class IcebergDlfRestCatalogTest {
         });
     }
 
-    private Catalog initIcebergDlfRestCatalog() {
+    private Catalog initIcebergGlueRestCatalog() {
         Map<String, String> options = Maps.newHashMap();
         options.put(CatalogUtil.ICEBERG_CATALOG_TYPE, CatalogUtil.ICEBERG_CATALOG_TYPE_REST);
-        options.put(CatalogProperties.URI, DLF_REST_URI);
-        options.put(CatalogProperties.WAREHOUSE_LOCATION, DLF_WAREHOUSE);
+        options.put(CatalogProperties.URI, "https://glue.ap-east-1.amazonaws.com/iceberg");
+        options.put(CatalogProperties.WAREHOUSE_LOCATION,
+                awsAccountId + ":s3tablescatalog/s3-table-bucket-hk-glue-test");
         // remove this endpoint prop, or, add https://
+        options.put(S3FileIOProperties.ENDPOINT, "https://s3.ap-east-1.amazonaws.com");
         // must set:
         // software.amazon.awssdk.core.exception.SdkClientException: Unable to load region from any of the providers in
         // the chain software.amazon.awssdk.regions.providers.DefaultAwsRegionProviderChain@627ff1b8:
@@ -93,42 +88,21 @@ public class IcebergDlfRestCatalogTest {
         // software.amazon.awssdk.regions.providers.AwsProfileRegionProvider@2792b416: No region provided in profile:
         // default, software.amazon.awssdk.regions.providers.InstanceProfileRegionProvider@5cff6b74:
         // Unable to contact EC2 metadata service.]
-        // options.put(AwsClientProperties.CLIENT_REGION, "cn-beijing");
+        options.put(AwsClientProperties.CLIENT_REGION, "ap-east-1");
         // Forbidden: {"message":"Missing Authentication Token"}
         options.put("rest.sigv4-enabled", "true");
         // Forbidden: {"message":"Credential should be scoped to correct service: 'glue'. "}
-        options.put("rest.signing-name", DLF_SIGNING_NAME);
+        options.put("rest.signing-name", "glue");
         //  Forbidden: {"message":"The security token included in the request is invalid."}
-        options.put("rest.access-key-id", ACCESS_KEY_ID);
+        options.put("rest.access-key-id", awsAk);
         // Forbidden: {"message":"The request signature we calculated does not match the signature you provided.
         // Check your AWS Secret Access Key and signing method. Consult the service documentation for details."}
-        options.put("rest.secret-access-key", SECRET_ACCESS_KEY);
+        options.put("rest.secret-access-key", awsSk);
         // same as AwsClientProperties.CLIENT_REGION, "ap-east-1"
-        options.put("rest.signing-region", DLF_SIGNING_REGION);
-        //options.put("rest.auth.type", "sigv4");
-
+        options.put("rest.signing-region", "ap-east-1");
         // options.put("iceberg.catalog.warehouse", "<accountid>:s3tablescatalog/<table-bucket-name>");
         // 4. Build iceberg catalog
         Configuration conf = new Configuration();
-        return CatalogUtil.buildIcebergCatalog("dlf_test", options, conf);
+        return CatalogUtil.buildIcebergCatalog("glue_test", options, conf);
     }
-
-    /**
-     * CREATE CATALOG dlf PROPERTIES (
-     *     'type' = 'iceberg',
-     *     'warehouse' = 's3://warehouse',
-     *     'iceberg.catalog.type' = 'rest',
-     *     'iceberg.rest.uri' = '<dlf-rest-uri>',
-     *     'oss.endpoint' = 'https://oss-cn-beijing.aliyuncs.com',
-     *     'oss.access_key' = '<your-oss-access-key>',
-     *     'oss.secret_key' = '<your-oss-secret-key>',
-     *     'oss.region' = 'cn-beijing',
-     *     'iceberg.rest.sigv4-enabled' = 'true',
-     *     'iceberg.rest.signing-name' = 'DlfNext',
-     *     'iceberg.rest.signing-region' = 'cn-beijing',
-     *     'iceberg.rest.access-key-id' = '<your-access-key-id>',
-     *     'iceberg.rest.secret-access-key' = '<your-secret-access-key>',
-     *     'iceberg.rest.auth.type' = 'sigv4'
-     * );
-     */
 }
