+++ fe/fe-connector/fe-connector-iceberg/src/test/java/org/apache/doris/connector/iceberg/catalog/S3TablesTest.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.datasource.s3tables;
+package org.apache.doris.connector.iceberg.catalog;
 
 import org.apache.iceberg.FileScanTask;
 import org.apache.iceberg.Table;
